update package to remove extra test
