+++ src/test/java/com/kyndryl/cjot/sample/stocktrader/account/test/BrokenJmsServiceTest.java
@@ -13,13 +13,12 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.kyndryl.cjot.sample.stocktrader.account.test.test;
+package com.kyndryl.cjot.sample.stocktrader.account.test;
 
 import com.ibm.hybrid.cloud.sample.stocktrader.account.AccountService;
 import com.ibm.hybrid.cloud.sample.stocktrader.account.json.Account;
-import com.kyndryl.cjot.sample.stocktrader.account.test.test.amqp.AMQPTestResource;
-import com.kyndryl.cjot.sample.stocktrader.account.test.test.couchdb.CouchDBTestResource;
-import com.kyndryl.cjot.sample.stocktrader.account.test.test.jms.JmsTestProfile;
+import com.kyndryl.cjot.sample.stocktrader.account.test.couchdb.CouchDBTestResource;
+import com.kyndryl.cjot.sample.stocktrader.account.test.jms.BrokenJmsTestProfile;
 import io.quarkus.test.common.WithTestResource;
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
@@ -38,14 +37,14 @@ import org.junit.jupiter.api.Test;
 import static io.restassured.RestAssured.given;
 
 /**
- * These test cases interact with CouchDB and AMQP. They do put messages on a messaging provider
+ * These test cases interact with CouchDB and have a broken AMQP Connection. They do not put messages on a messaging provider
  */
 @QuarkusTest
-@WithTestResource(value = CouchDBTestResource.class, restrictToAnnotatedClass = true, parallel = true)
-@WithTestResource(value = AMQPTestResource.class, restrictToAnnotatedClass = true, parallel = true)
+@WithTestResource(value = CouchDBTestResource.class, parallel = true)
 @TestHTTPEndpoint(AccountService.class)
-@TestProfile(JmsTestProfile.class)
-public class JmsServiceTest extends AbstractIntegrationTest {
+@TestProfile(BrokenJmsTestProfile.class)
+public class BrokenJmsServiceTest extends AbstractIntegrationTest {
+
     @Test
     // Set up the JWT/Security items
     @TestSecurity(user = "stock", roles = "StockTrader")
@@ -57,7 +56,7 @@ public class JmsServiceTest extends AbstractIntegrationTest {
             @ConfigMetadata(key = "issuer", value = "http://stock-trader.ibm.com"),
             @ConfigMetadata(key = "audience", value = "stock-trader")
     })
-    public void jms_testPlatinumUpdateAccountEndpoint() {
+    public void jms_testPlatinumUpdateAccountEndpoint()  {
 
         Account account = new Account(faker.name().fullName());
 
@@ -78,6 +77,7 @@ public class JmsServiceTest extends AbstractIntegrationTest {
                         .extract().as(new TypeRef<>() {
                         }); // return the values back
 
+        // We should still get good data back.
         // Verify everything that we created was returned.
         Assertions.assertEquals(account.getId(), persistedAccount.getId());
         Assertions.assertEquals("Platinum", persistedAccount.getLoyalty());
