+++ src/test/java/com/kyndryl/cjot/sample/stocktrader/account/test/CouchDbReadinessTest.java
@@ -13,9 +13,9 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.kyndryl.cjot.sample.stocktrader.account.test.test;
+package com.kyndryl.cjot.sample.stocktrader.account.test;
 
-import com.kyndryl.cjot.sample.stocktrader.account.test.test.couchdb.CouchDBTestResource;
+import com.kyndryl.cjot.sample.stocktrader.account.test.couchdb.CouchDBTestResource;
 import io.quarkus.test.common.WithTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.http.ContentType;
