+++ src/test/java/com/kyndryl/cjot/sample/stocktrader/account/test/NoCouchDBAccountServiceTest.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.kyndryl.cjot.sample.stocktrader.account.test.test;
+package com.kyndryl.cjot.sample.stocktrader.account.test;
 
 import com.ibm.hybrid.cloud.sample.stocktrader.account.AccountService;
 import com.ibm.hybrid.cloud.sample.stocktrader.account.json.Account;
