+++ src/test/java/com/kyndryl/cjot/sample/stocktrader/account/test/NoCouchDbReadinessTest.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.kyndryl.cjot.sample.stocktrader.account.test.test;
+package com.kyndryl.cjot.sample.stocktrader.account.test;
 
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.http.ContentType;
