+++ src/test/java/com/kyndryl/cjot/sample/stocktrader/account/test/amqp/AMQPTestResource.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.kyndryl.cjot.sample.stocktrader.account.test.test.amqp;
+package com.kyndryl.cjot.sample.stocktrader.account.test.amqp;
 
 
 import io.quarkus.test.common.DevServicesContext;
